+++ trunk/xmipp/libraries/reconstruction/angular_continuous_assign.h
@@ -84,7 +84,7 @@ public:
     /** Predict angles and shift.
         At the input the pose parameters must have an initial guess of the
         parameters. At the output they have the estimated pose.*/
-    void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId);
+    void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId);
 
     /** Finish processing.
         Close all output files. */
